feat: Allow setting version via Env var
Use case is where we're already creating releases via sentry-cli, but we want the post-publish hook (as getting sourcemaps otherwise seem to be an issue and this was made to resolve it judging from https://github.com/expo/expo-cli/issues/298). Creating a new release with an existing version via sentry-cli seems to be a no-op.