Only disable Sentry in bare workflow in development explicitly.
It's possible not to set the `enableInExpoDevelopment: true` flag, and then upload/release a build on the appstore and have Sentry be disabled if the build wasn't "published" or "exported" yet.

This only disables Sentry in development if that flag is actually false, and enables Sentry by default in production, unless explictly disabled.

.